requirements.txt
discord.py
asyncio
mcstatus
youtube_dl
ffmpeg
pynacl